experiments: introduce ExperimentXLabManager
Add a new manager somewhat compatible with the existing
ExperimentUserManager. The manager will be used when
MetricsPlatform is available and the flag
GEUseMetricsPlatformExtension enabled.

Also add it as JS config var for the Growth client to
know from which source to pick up the variant assigned
while two bucketing systems coexist.

The manager holds a list of valid experiments to
use it as an allow list for experiments the extension
can handle. Since GrowthExperiment code can only run
with a single experiment-variant assgined at the same time,
if two enabled valid experiments are found, the manager
will return the assigned variant for the first
enrolled experiment, meaning the user is sampled.

Given the following configuration:
```
[
	'name' => 'growthexperiments-some-name',
	'groups' => [
		'control',
		'treatment',
	],
	'sample' => [
		'rate' => 0.5,
	],
],
```
users will be assigned as:
 - Unsampled users get the "control" variant defined in
GEDefaultHomepageVariant
 - Sampled users in control get the
"growthexperiments-some-name_control" variant
 - Sampled users in treatment get the
"growthexperiments-some-name_treatment" variant

Bug: T375525
Bug: T401308
Change-Id: I753f96dcbc25b6df93a5311a1bf3cf1efa938299