MentorStore::hasAnyMentees: Use more efficient implementation
We currently fetch all mentees and then check whether the result
is empty or not. This is a super inefficient way to find out if
a mentor has any mentees. This patch rewrites this to use
SELECT ... LIMIT 1 as the underlying query, which would be much
faster to compute.

This is a BREAKING CHANGE. Since this patch, the $includeHiddenUsers
parameter for MentorStore::hasAnyMentees is ignored and always assumed
to be true. As far as I could see, this capability is unused, hence removal
makes sense, but I'm keeping the parameter in (hard-deprecated) just in case,
as well to make the intended effect of this patch more visible and easier
to review.

Bug: T376124
Change-Id: I08bff4ecdf4f19ad34cc83d39c20f52be67bc42e