ReassignMenteesJob: Improve logging
This patch does two things. Firstly, it fixes
incorrect logging in the job itself. Previously,
the job claimed all mentees were reassigned if it
did not finish successfully. This patch skips the logging
entriely in that scenario.

In addition to that, it adds logging for failure to release
a lock. This is one of the places where things might be breaking
that would cause the lock to not be released before the new job
starts.

Bug: T376124
Change-Id: I1e242112b4544d6f6b639dbf0d815b2007c228ff