refactor(tests): don't use per-method coverage annotation
Coverage annotation in the doc-block of the class is the recommended
approach (see https://docs.phpunit.de/en/10.5/annotations.html#covers).
Per-test method-wise coverage annotation has the problem that it either
does not cover private methods or the test needs to know about private
methods which are supposed to be a hidden implementation detail.

Depends-On: Ie76020dc4fa3545f827e1674051530b479f01f31
Change-Id: If6657f02e33f36860fd27b3f326f5eacdf92afc3