ExperimentsUserDefaultsManager: introduce user bucketing condition
Add a ConditionaUserOption condition CUCOND_BUCKET_BY_USER_ID to
assign a variant to all users which do not have one explictly stored
in growthexperiments-homepage-variant based on hash of their user
centralId. This allows to assign a variant to existing account holders
without requiring to store it in the DB.

An example of expected wgConditionalUserOptions configuration is:
```
$wgConditionalUserOptions['growthexperiments-homepage-variant'] = [
 [ 'bucketA',
	[
		ExperimentUserDefaultsManager::CUCOND_BUCKET_BY_USER_ID,
		'experiment',
		50
	]
],
[ 'bucketB',
	[
		ExperimentUserDefaultsManager::CUCOND_BUCKET_BY_USER_ID,
		'experiment',
		100
	]
],
]
```
Where the default value (eg. bucketA) is the variant to assign,
and the arguments to the CUCOND_BUCKET_BY_USER_ID condition are
a name for the experiment and an integer representing the accumulated
probability for the variant to be assigned. Noted as a percentage (0,100).

Note: Both ExperimentsHooks and ExperimentUserDefaultsManager
are introduce to circumvent circular dependencies between
user options lookups, condtional defaults, and centralId.

VariantHooks will still run onGetDefaultOptions setting a default
variant named "control" as the default user experience if there
are no experiments set in wgConditionalUserOptions.

Bug: T376266
Depends-On: Ife6f96397eafd61fdb40528aac315ddde1ef2774
Change-Id: I4bf8cec6de4082ef3240a1f3fa3848eda04bc008