CommunityUpdates: Improve error handling
CommunityUpdates attempts to make an API call,
and silences any possible errors that can happen
there (response is not a JSON, API is unreachable, etc.).

At the same time, it throws an exception when API URL
is not configured, while this situation can be gracefully
handled by returning no thumbnail (similar to what we already
do when the API errors out).

What:
* Do not throw when API is not configured
* Use Util::getJsonUrl rather than the request factory directly,
  as the Util method already does most of the error handling.

Change-Id: Ie5526d2b61814673d974df75e5517c78a5483443