Explicitly start transaction in maintenance script
Before this change being merged, running the maintenance script to prune
dangling database-records would cause a lot of errors in logstash
complaining that there is no transaction to commit, while still actually
executing the write to the database and removing the dangling records.

The error message was as follows:

```
GrowthExperiments\Maintenance\FixLinkRecommendationData::fixDatabaseTable: no transaction to commit, something got out of sync
```

The line to commit the transaction stems from the original version of
this script as it was merged 3 years ago in
I861db8702b30c4559a19c1bdd6d108d12389fc23. As far as I can tell, there
was never an explicit begin of the transaction. It is unclear whether
there ever was an implicit begin of the transaction, or if maybe only
the way we log errors for unmatched commits of transactions was changed.

Change-Id: I72543e7e3b0bdedf098aa155d53a257c966072e9