Help panel: Disregard namespace config when in Suggested Edits mode
Previously, Help panel was not getting enabled in Suggested edits
mode when it was disabled via Community Configuration. This is
not desirable, as being in Suggested Edits mode should override
all user-specific and site-specific configuration,
except the server-side feature flags.

This was allowing admins to break Suggested Edits by disabling
the Help panel in the mainspace.

Bug: T377862
Change-Id: If5a9aed0069f85494f664d9bc303732c2c9ed045