Create no-link-recommendation variant
Why:
We want to deploy Add Link to English Wikipedia in a more
controlled manner, to avoid overwhelming users with new edits.
This can be done with ExperimentsUserManager's help.

Since link-recommendation are normally available, and we have
to keep them available in `control` group (default experience),
this patch creates a `no-link-recommendation` variant to disable
Add Link on per-user basis.

What:
* Introduce `no-link-recommendation` variant
* Variant can be enabled only manually via developer tools
  (ge.utils.setUserVariant).

Bug: T377787
Change-Id: I3a56c07cd72e08bbaf086b899eac3b9d255b25dd