maint: fix stats-collection flakyness by migrating to statslib
These two metrics that we collect from maintenance scripts have a
history of being very spotty with a lot of events missing and there not
being an error message to account for that. The hope is that the new
stats-system is more robust and will result in more reliable collection,
transmission and storage of data.

The main focus of this work is to improve the reliability of our
collected data in accordance with our current annual goals and quarterly
hypothesis. As a secondary effect, it is also contributing to our
overall migration to statslib.

Bug: T372337
Bug: T374050
Change-Id: I14701a0a6d8e6e93770257e1b82db9b3442bb8d0