[i18n] mentorship: Make `toohigh` error message more generic
Awayness of the mentors can be changed in two places:
* by the mentor at Special:MentorDashboard,
* by an admin at Special:ManageMentors

Unfortunately, the field is called differently in both cases.
At Mentor dashboard, it is called `You'll be away for`, while
at Special:ManageMentors, it is called `Away until`.

As the error message is shared in both cases,
make its text a little bit more generic by
not referencing the field by its name.

This change was approved by Kirsten at the attached task.

Bug: T379474
Change-Id: I785f299f03859669b81ab20820722e7983e8a087