AwaySettingsDialog: Do not parse error message on the client side
On an error, the API provides the caller with two pieces of information:

* an error code (in practice, an i18n message)
* a human-readable description of the error (in practice, the text
  of the i18n message)

AwaySettingsDialog was using the error code, parsing it on the client
side using `mw.msg`. This did not work well, as messages might have
parameters the client is unaware of.

Since we already have the parsed version of the message, let's use
it directly and avoid the parsing. This is in line with how API error
handling is done in other similar places in GE (MentorMessageChangeDialog,
HelpPanelProcessDialog).

Bug: T380149
Change-Id: Id91ca20fed4b872c8ea8e61fd6fb8154f817c772