SearchStrategy: make one query for ORES topics filters
Why: performing a single query per topic was implemented in
T243477 to circumvent query length limitations. This approach leads
to count double articles that are tagged with more than one
articletopic preent in the query.

What: since the query limit has significantly grow since the
initial implementation, and making less queries should be
more performant, change the search strategy to build a single
query when all topics in the filters are ORES. Mixed topics
in filters are still supported but use the old strategy
of query-per-topic.

Bug: T373188
Change-Id: I6ffce14359148bee5e63a2acb4bab99f2491799c