AwaySettingsDialog: Check range on the client side
AwaySettingsDialog is actually already aware of the 365 days limit
(it is hardcoded as the HTML max attribute). This means we can
check the input's validity on the client side (without involving
the API).

This has several benefits:

    (1) Quicker response for the user, as we do not need to wait
        on the API.
    (2) Works for any input time (JS is unable to generate ISO
        timestamps for dates that are in far distant future; this way,
	we can only ask for a timestamp if the value is actually within
	range).

As a bonus, this patch moves the 365 value to a constant,
which should make its presence more obvious to future
developers (in addition to being generally the nice thing to do).

Bug: T380149
Change-Id: I936226a0a1c03a4fcb2e9bd09cfecff7ff205361