fix(NewcomerTaskStore): ensure preloaded first task matches current filters
Why: the store adds the preloadedFirstTask to the task count after any
request is made. While that may be correct for some where the enabled
filters at loading exactly match the ones requested, it is wrong for
all requests where there are filter differences.

Wht: check filters have not changed before adding the preloaded
first task to the queue and count

Bug: T369806
Change-Id: I0d35517bab5838361ff7b66e758a014b7796201c