feat(surfacing): track performance metrics
These are aggregate metrics to be able to tell how our feature is doing.

The "scroll-away" metrics are tricky, because that intersection observer
fires every time that popup is no longer visible, including when the
user clicks any buttons in it or outside. This is fine for the purpose
of closing the popup, but when tracking then we actually want to know
whether the popup disappeared because the user scrolled away or because
they clicked "No".
Let's leave that rabbit hole for later.

Bug: T377097
Change-Id: I29ad7f546e9fafee81f50f63e52958bf92b34b6c