test(cypress): fail if there is an error in the console
We need to keep track of all the spies we create in a test, in case we
navigate to multiple pages during a test.

The `trackSurfacing` method is not available right away when a page is
loaded, but only after resource loader is fully initialized. Thus we
keep waiting until it is available. Sometimes that means just waiting
for the already queued async actions (thus the `0` initial timeout), but
sometimes a bit more time is needed and thus we are retrying in 100ms
increments.
Overall, it might make sense to move that logic into a dedicated file or
to upstream it somewhere, given that it is largely independent of
GrowthExperiments.

The changes to the LocalSettings are necessary because there actually
was an error in the Impact-test coming from the suggested edits module,
complaining that the TaskSuggesterFactory was not correctly set up. This
now adds a stub for that to this change. Follow-up changes will fill
that factory with proper testing for migrating the existing add-a-link
and add-a-image and template-based task changes.

Bug: T377198
Change-Id: If651480d6a928d710c00a810c809a37bfc80b2da