test(cypress): work around Chai "feature"
Chai, a Cypress dependency has a "feature" that to pretty-print an
object it tries to invoke the `inspect` method on that object if it
exists, expecting to receive a meaningful string representation.
Obviously, this harshly clashes with any other uses a method called
"inspect" might have on an object.

In our case, the `mw` global js object has an `inspect` method to report
module performance or something. However, this method being invoked with
the entirely unrelated arguments provided by Chai is leading to a
ResourceLoader exception.

That is triggered when referencing `mw` in a
call to `.its` from `cy.window()`. This change works around that problem
by referencing `mw.loader` instead and checking for it having the
`using` property, by which point the RL modules should be loaded and we
can make API calls.

For the upstream cypress/chai bug see: cypress-io/cypress#30662

Change-Id: I4bc0c39c8f50cb96eccf409bdcab4554b84af529