fix: show thumbnails in surfacing popups
Per https://www.mediawiki.org/wiki/Codex#Usage_without_JavaScript_(CSS-only_Codex_components)
we need to explicitly list the styles to be available for our CSS-only
components.

It is unclear why that was not an issue locally. Is something else
requiring that component or its styles or codex as a whole?

Bug: T381364
Change-Id: I17fa13127d2204a3f5fd4fac2553be2bafd63cc7