refactor(Mentorship): remove noop code
This was introduced in I21d75c6c (commit 2ea8878) for T264343 and was
used in the class `WikitextMentorProvider`. However, that class was
removed in I5e3cfb120 (commit 75bd820) for T321501. Leaving this code in
place was likely just an oversight.

It is now being removed in the course of migrating this hook handler to
domain events. (See T380498)

Change-Id: Ie84dfa7e4a35c6e21d3b7b94f0d2e8f16462cd91