ExperimentUserDefaultsManager: introduce CUCOND_BUCKET_BY_LOCAL_USER_ID
In order to circumvent outdated data issues at the time
of account creation in CentralAuth (T380500), create a
bucketing condition that uses local IDs as opposed to
central's so variant assigned can be tracked at the time
of account creation.

This has the side effect of users getting different
variant assigned in different wikis but that is
acceptable since Growth features are first rolled out
to small set of (pilot) wikis, then scaled to all.

Bug: T379682
Change-Id: I94d329466be30920572515c359ab4dbf71cb64d0