Remove legacy way of clearing link recommendations + temp config
With the new config, the data is consistently good, so the new hook
seems to be working as expected. We noticed that the
revalidateLinkRecommendations maintenance script is still add new
dangling records, but that is a separate issue.

See I3b941c46422500746c73c1aa9f21fbcbd01bbbec for the change removing
this from wmf-config.

Bug: T379522
Change-Id: Id70d05b05ebd5d8a1650208b28b435da1f89d49e