[tests] Migrate ComputedUserImpactLookupTest to Community Configuration 2.0
CommunityConfiguration extension prepopulates the configuration
with defaults according to the schema. Because of that,
the empty test now works with newcomer tasks defined, while
previously it used to run with no tasks defined.

Fix this by changing the assertion to match the newly-generated
zeros. We also switch to assertEqualsCanonicalizing, as the array
order is irrelevant (and not negligible).

Alternatively, we could change how the feature works (by
not incuding zeros in the result), but that seems like a wider change.
We can also remove the data with zeros and compare with an empty
array, but that seems like more logic in tests than necessary.

Bug: T380580
Change-Id: I71c770a77f58af266aebaf6dacab956de5cbdd29