chore: QuestionPoster: Add typehints
This amends the constructor and property typehints.

== Constructor change
The class is only created from QuestionPosterFactory, which guarantees
(via a type hint) that $body and $relevantTitleRaw is a string.
Hence, we can have the same guarantee in the constructor.

== $relevantTitle
$relevantTitle is documented to be Title|null, but it might not be
initializated in all codepaths. Since it was not typehinted, accessing
uninitializated property worked (returning null, as expected). However,
typehinted properties cannot be accessed before initialization. To
remedy this, set the property to null (as it is lazy loaded
in getRelevantTitle()).

Change-Id: I9b034d42bbc16726d8163108d457b897cdef1cd3