fix(surfacing): prevent console error if no recommendations found
These "errors" are benign as they are the expected reasons for a request
to return no recommendations, like it has found not enough good
recommendations or the last edit was already a link-recommendation one.

However, adding this information like this causes the api-client in the
browser to throw an exception and we would need to handle that
explicitly there. That seems undesirable, especially since these reasons
are not intended for the users in any way and are only interesting to
engineers. Thus, it seems desirable to just not send them with the
response, also making it ever so slightly lighter.

Bug: T378536
Change-Id: I3af86e73a306e96b17cca5d052afbe114ff725b7