fix(tracking): TimingMetric:observe records milliseconds
It turns out, see T355837#10438043 and following comments, that it now
actually matters in what unit the timings are being provided. If it is
the wrong unit, the recorded data is essentially useless.

We also cannot use `copyToStatsdAt`, because that would essentially
multiply our existing metrics by 1000, see T359248#9994136 for this
happening.

It is sadly necessary to inject the dependency into
LocalSearchTaskSuggester because otherwise its unittest fails.

Bug: T383208
Change-Id: I8e80de1c99f476acde290c7d442d05f5c62c369d
Depends-On: I0c56cc8153318dc79b94a8bc390b7dfabcf8ca38