Migrate ORES topic map to PHP configuration
Previously, the ORES topic map into Growth topics
was stored at MediaWiki.org as MediaWiki:NewcomerTopicsOres.json,
loaded via legacy Community Configuration.

To enable us to remove legacy Community Configuration
code from the repository, this patch moves the configuration
elsewhere. I decided to put it as a config variable in extension.json,
as it is not something we want to allow admins to modify.

Note this patch leaves config for `morelike` intact, as that is
a legacy way to access topics, which isn't going to be enabled
in practice.

This patch also switches the default topic system to `ores`.

Bug: T367576
Change-Id: I72ec95c3bdd631e19aa5996653f41fffab2ac81d