ConfigHooks: Do not unconditionally depend on legacy CC
In order to deprecate legacy CommunityConfiguration, we need
to drop all places where we unconditionally depend on it.
Unfortunately, one of those places is hook handlers.

It doesn't seem to be possible to unconditionally register
hook handlers. To workaround that, this patch creates an empty
ConfigHooks implementation. When the handler is constructed,
we check if CommunityConfiguration extension is enabled,
and return the placeholder implementation or the legacy implementation,
as appropriate.

Bug: T383330
Change-Id: I3c6c1ed4e63ffb909aa075fe288319c121af6352