tests(Cypress): migrate-selenium test for suggested edits navigation on homepage
This is laying the groundwork for migrating the more complex tests
link-suggestions and image-suggestions to Cypress (and CC2.0).

After considering it more, I decided to keep the pattern of a
PageObjectModel for the homepage because this page is sufficiently
complex and the identifiers we use are often so obscure that this makes
sense to me.

This is partly a departure from a (very opinionated) blog
article by Cypress:
https://www.cypress.io/blog/stop-using-page-objects-and-start-using-app-actions
However, that article is mainly about _setting up_ the environment for
test, less about the actual test. We still use stable interfaces where
possible for setting up the test-specific environment (maintenance
scripts, mw.Api for login and setting user options, etc.).

Also, the discussion here articulated well some of the concerns I had about
that overall approach:
https://old.reddit.com/r/softwaretesting/comments/y1j8rq/should_i_use_page_objects_with_cypress/

The pages are extracted from
tests/selenium/fixtures/SuggestedEditsContent.xml

The `DecoratingTaskSuggesterFactory` has been added, because what is
currently in the selenium-version of the thing is causing an error in
the javascript console in the browser. Selenium doesn't care about that,
but we are intentionally failing Cypress tests on errors in the console.
Thus setting it up in a way that is a bit closer to production and thus
fulfilling the expectations of other code.

Bug: T380581
Change-Id: I608484c5d0565fafea14d4f777e820d0f72d53ad