tests(Cypress): support multiple tests per file
Most importantly, this change moves the
`
Cypress.on( 'window:before:load', ( win ) => {
`
event handlers out of the `beforeEach` event listener, because those
`Cypress.on(` listeners persist for the entire test-suite (i.e. file)
and so should only be executed once for each. Otherwise we would get
errors like `Trying to wrap X but it is already wrapped` or something
similar when we have multiple tests per file.

Wrapping this in a `failOnConsoleError` function just makes this
files structure a bit clearer, that's all.

Change-Id: I74ece7afaab2caa687f82c7f2e7284f7eedb8f04