Config: remove migrateCommunityConfig
Why: The script was used to migrate existing config in
WMF hosted wikis from CC version 1 to 2. The script
was released in mw 1.43 for allowing wmf and non-wmf
wikis to do the migration. Since it takes in account
the schema state at the moment of the migration,
May 2024 any later changes to the schemas can break
assumptions and perform a buggy migration.

What: remove the script as it wouldnt be longer
maintained

Change-Id: I5de09ccf3299e70e223fd272a20ebebbf7afe874