tests(Cypress): migrate Selenium addlink to CC2.0 and Cypress
This removes the add-a-link related parts from the Selenium directory,
which leaves us with no selenium tests at all. But to keep this diff
small, the entire selenium directory and all related code is cleaned up
not in in a follow-up commit.
However, we must remove the command to run the selenium tests in
package.json, because it fails when there no specs at all.

Bug: T380581
Change-Id: Iaf9530d92d863334e375042c078b058c694e3e86