fix: resolve race-condition on pcs-request failure
This race-condition would cause an error when
1) the request from pcs would fail
AND
2) the user clicked forward past the end of the queue ("the no more
   tasks"-screen) before the failed pcs-request returned.

This is mostly prone to happen in the CI browser testing when asserting
the correct behavior of the queue.

This is fixed by using the same task object in the fail branches that is
also modified by the request methods themselves. Also, renamed that
object to make its identity and purpose clearer.

Change-Id: I212e566eeeba57486b2973c40d55ce2b7c511629