tests(Cypress): don't fail test on subscribeWhenAvailable failure
On local occurrences of this error, win.document still existed. So it is
unclear what might be causing this, but it is probably not worth that
this increases the flakyness of all our Cypress tests.

This still logs a warning to the browser console which should show up if
a test fails one of its assertions. This might highlight if the
inability to call `mw.trackSubscribe` is related to some more fundamental
issues affecting other parts of the page as well.

Bug: T383441
Change-Id: Icbe74359665699856806c26902b54a23d5280e64