fix: track correct data in legacy Graphite query
This is a bug that was introduced in ee8695f. In that commit, the statslib call was
tracking the correct variable, but the legacy Graphite/statsd one
was copied with the metric name from line 114. This change fixes that so
that the statslib call and the legacy statsd call are both tracking the
same data.

Bug: T383208
Change-Id: Ic62260994422d260ca6f74a97652a738d802391d