tests(Cypress): 2nd try to fix "center of element is hidden"-error
The Template-based tasks cypress test is still flaky with on the last test the
typing failing with:

```
CypressError: Timed out retrying after 20000ms: `cy.type()` failed because the center of this element is hidden from view:

`<div class="ve-ce-branchNode ve-ce-documentNode ve-ce-attachedRootNode ve-ce-rootNode mw-content-ltr mw-parser-output mw-show-empty-elt" contenteditable="true" spellcheck="true" tabindex="0" data-gramm="false" role="textbox" lang="en" dir="ltr">...</div>`

```

Previously, we tried in a9d0ae0 to fix that by making the article
shorter. However, that did not result in the desired success.

Now this change tries to fix the issue by adding
`{ force: true }` to the type command.

In Ic5d7847389, we ran the core of this test in a loop many times, and
while it always failed eventually, it was never with this error itself
and never when trying to type into a page. Rather, the error seemed to
stem from running in a loop itself (saving a page took longer and longer
until it timed out.)

Further changes: this hides per user-config the VE "Welcome to
test-wiki" banner that caused problems before.

It also moves the assertion that the help-panel does not exist. The
reason is that the assertion directly after clicking the "edit"-button
would pass even if the help-panel would open, because it would run
_before_ the helppanel would have been attached to the DOM. (That is no
problem for closing the help-panel, because that waits until the
help-panel exists.)

Change-Id: I588165ed3ce52ebd9a46dd85837be9fbd49d3ce7