feat(surfacing): show only the clicked suggestion
Add a query parameter `recommendationid` to track
which recommendation did user click in reading mode.
Make use of the `link_index` property returned by the
API to identify recommendations.

The results and order should match the ones added in
the suggested edit session before any sorting or filtering.
That means the index can be higher than the recommendations
shown for an article which are cap by maxLinks config
setting.

Bug: T381450
Change-Id: I6854561cdc4a725adced12e22da10c8e584071cd