Make dock button compatible with Minerva code
In I76dfc06d3529d00f393cd618ea4a815d9531803f p-dock-bottom will
be added to the page. This will disrupt the styling of the icon.

This change makes sure that the styling is restricted to when
the button is a direct child of body element rather than

See mw.util.addPortletLink call in HelpPanelCta.js

Bug: T384733
Change-Id: I3a46354bb69d5c4b17b718f96df3f7d84a649956