i18n: delete messages for article topics being moved to WikimediaMessages
All these messages are being moved to the WikimediaMessages extension,
for easier reuse. The first step of the migration plan, as described in
T380825#10481384, is to move all message definitions verbatim, i.e.
remove all messages from here and add them to WikimediaMessages with no
other changes (see patches with the same topic, which should be merged
together with this one).

The `-history-and-society` and `-science-technology-and-math` are left
untouched because the ContentTranslation version will be used instead.

The `-campaign` message is also untouched because the group does not
exist in the new taxonomy.

Bug: T380825
Depends-On: Ic741bf5b5e52365c7a7f09330b6ae694b61b028a
Change-Id: If5c29bf28cd09c663c5fac2007dd7ceeaa01e440