refactor(AddLink): Make eval steps more legible
The goal of this change is to lay the ground-work to being able to react
more meaningful to a page not having link-recommendations available for
various reasons. The follow-up change I2ea572ed8 will make us of this to
store specifically when a page has no recommendations for reasons that
are likely not to change until the page is edited again or the
configuration changes.

Another follow-up change is planned to build on this to track more
metrics when searching for new link-recommendations, so that we
understand better why we have the recommendations that we have and not
more.

The reason for introducing a new method over modifying `get` directly,
is that `get` is also used in all the ImageRecommendation providers.
That means modifying that interface would also mean modifying all of
them.

Change-Id: Ie4b5ab43a48e8227d2ffec205e4a131c27a2dafb
(cherry picked from commit 5821575b78e0f636f434bddd33bc561269d12591)