fix(surfacing): adjust styles to closer match VisualEditor ones
This uses design-tokens to match 1) support dark-mode on desktop and 2)
match the general guidelines for structured task highlights from
T380285. This is also the reason for the dashed border.

In addition, this adds more spacing (margin+padding) to match the
spacing we have for these highlights in VisualEditor. This spacing issue
has been especially pronounced in Chromium based browsers.

Bug: T385764
Change-Id: I285e9ac85c315bdba47482997a9b9d5bdc1f6951