refactor(AddLink): ignore rows with `null` in Store
This small change is intended to be effectively a no-op because there
are no rows with null yet. It exists so that it can be packported to
the current production branch, so that it would be easier to roll the
change back that actually inserts rows with `null` without the database
subsequently breaking.

Bug: T382270
Change-Id: I4250bcc676b3e04d8275a01eaf338d1525c4944f
(cherry picked from commit a48df00a61ae90232b91e68796181456b4995081)