Mentorship: Make the dropMenteeRelationship call useful
The previous code called dropMenteeRelationship() if
$mentorUser was falsy. This is incorrect; when $mentorUser
is falsy, the relationship does not exist, and there is nothing
to drop. This means the dropping was no-op since its introduction.

The purpose of the drop was to do a cleanup, so this did not result
in any user-facing error, but we have relationships stored that we
actually do not want. With this patch, they should be dropped slowly.

Add a test to prevent a similar issue from happening again in the future.

Change-Id: I61be14caed2de97af8af5bcb2c33e86562636155