i18n(surfacing): update fallback copy to singular suggestion
In context of T381450 we changed it so that only ever the clicked
suggestion is presented to the user when they get to VE. The copy should
reflect that.

That copy is used when for some reason more information about the
article can not be loaded. Usually instead of this copy, a link to the
article with an image and description will be shown.

Bug: T386236
Depends-On: Ieb9c5856a5cbb07992779dee22ec6adc92fe12b8
Change-Id: Ice51b2ace968217b7441459058ef4c9339a4bf6e