Disable Newcomer tasks and Link Recommendations if dependencies are not met
Newcomer tasks and in particular Link Recommendations depend on further
extensions being installed. Provide utility methods to check for these
preconditions.

Unfortunately, that does not mean that we can drop VisualEditor as hard
requirement from the extension.json just yet. The reason is that we
depend on VisualEditor-hooks in extension.json and so the respective
interfaces are required to exist.

Bug: T386690
Change-Id: I5efefb6d86807c4bf721ed2bf16808e20e9ccf81