refactor(AddLink): use smaller interfaces
This will streamline future work when we no longer have to build full
Title objects. For now, we still do that in the initial step to check
for existance, but that can probably also be done in a better way.

PageReference::__toString() is explicitly documented as being suitable
for use as a cache-key, as well as logging and debugging. A cursory
glance at the implementations seems to confirm this. Thus using it in
that capacity here.

Change-Id: I1d37dcfe4d66d039301ca03d56be772d4aafe6d1