fix: only depend on CirrusSearch if image suggestions are enabled
This is another hack to not add a hard dependency on CirrusSearch to
GrowthExperiments. The proper solution here as well would be to factor
out the entire "newcomer tasks" functionality into its own extension and
have that depend on CirrusSearch (and VE and WikimediaMessages) in a
hard way.

With this change, the `GrowthExperimentsAddImageSubmissionHandler`
service can be constructed even if CirrusSearch is not available. An
error will only be thrown in that case if the service is actually used.

Bug: T386690
Change-Id: I7929ffe102b9ed96da4b924a4a2005330509ea2c