Obtain article topic definitions from WikimediaMessages
In order to allow the migration in T380825 to move forwards, rely on the
`ArticleTopicFiltersRegistry` class provided by the WikimediaMessages
extension to get topic messages for ORES-based topics. The base `Topic`
class is left unchanged because it seems like its scope extends beyond
the predefined list of topics; and the "Campaign" topic also gets
method overrides to keep using the group name defined here.

For the time being, topic definition themselves are also not really
taken from WikimediaMessages. Instead, we keep using the config setting,
and enforce a validity check on group and topic names. So:
- When loading the config value, make sure that all topics are
  recognized by WikimediaMessages. If not, show a user-visible error.
  Note that this currently uses a hardcoded English message, the reason
  being that the fix is presumably only temporary™.
- In OresBasedTopic, assume that the given topic and group IDs are
  valid, and hold the caller accountable for validating them; mention
  this in the constructor documentation.
- Update the existing PHPUnit test to use a valid config baseline. Add
  a new test for unrecognized topics (separated from the existing test
  due to RawMessage special handling).

Avoid hard depending on WikimediaMessages by skipping topic loading
when WM is not available and return an empty array of topics as fallback.
Also make WM a dependency for enabling topics in the client via the js
config var GEHomepageSuggestedEditsEnableTopics.

A future change might remove the option to configure article topics via
the extension config, but the goal of this patch is to unblock further
work in T380825.

Bug: T386018
Co-Authored-by: Sergio Gimeno <sgimeno@wikimedia.org>
Change-Id: Ie05689c803098bda461a6de26fb17b49f41239bb