tests(Surfacing): Reenable and extend cypress test
The stubbed version of this test was originally created back when we had
not yet any scaffolding for Cypress in place. With the migration of all
Selenium tests to Cypress, this scaffolding now exists and we can use it
to properly complete this test.

This test was disabled when adding the condition that
link-recommendations must exist for a page in I415859f25. That was no
longer compatible with stubbing the api-responses because it affected
the initial page-load.

The new test now also asserts that the expect new tag for edits from
surfaced suggestions is added to the revision.

Follow-Up-To: I415859f2559d118f59b5f2079d1a216fe598ab57
Change-Id: Ied65a8a9acc6147457db8782b293e49770f38b81