refactor(refreshLinkRecommendations): extract method iterating over ores-topics
This is a pure noop refactoring of just extracting that code into a
separate method. It is very self-contained, so $force is the only
parameter we need.

This is done in preparation for the next change that adds the new way of
refreshing link recommendations by iterating over all pages.

Change-Id: I5a74e5ffbbf4f25f1f8ea8b8c621e5a13f01ea03